// api/order.entity.ts
export class Order {

  id: number = null;

  productId: number = null;

  paymentId: number = null;

  shippingId: number = null;

  userId: number = null;

  amount: number = null;

  type: string = 'default';

  email: string = null;

  salutation: string = null;

  title: string = null;

  firstName: string = null;

  lastName: string = null;

  tel: string = null;

  fax: string = null;

  mobile: string = null;

  country: string = null;

  state: string = null;

  city: string = null;

  zip: string = null;

  street: string = null;

  streetAdditional: string = null;

  number: string = null;

  message: any = null;

  files: any = null;
}